优化a 弃用b

+++ docker-entrypoint.sh
@@ -69,31 +69,22 @@ if [[ -d $ICQQ_PLUGIN_PATH"/.git" ]]; then
     else
         git pull origin main --allow-unrelated-histories
     fi
-    if [[ -e $ICQQ_PLUGIN_PATH"/config.yaml" ]]; then
-        if [[ ! -e "/app/Yunzai-Bot/config/ICQQ.yaml" ]]; then
-            echo -e "检测到 ICQQ-Plugin 配置文件旧位置，复制到config/ICQQ.yaml\n"
-            cp config.yaml /app/Yunzai-Bot/config/ICQQ.yaml
-            mv config.yaml ICQQ.yaml.bak
-        fi
+    if [ -e $ICQQ_PLUGIN_PATH"/config.yaml" ] && [ ! -e "/app/Yunzai-Bot/config/ICQQ.yaml" ]; then
+        echo -e "检测到 ICQQ-Plugin 配置文件旧位置，复制到config/ICQQ.yaml\n"
+        cp config.yaml /app/Yunzai-Bot/config/ICQQ.yaml
+        mv config.yaml ICQQ.yaml.bak
     else
-        if [[ -e "ICQQ.yaml.bak" ]]; then
+        if [ -e "ICQQ.yaml.bak" ] && [ -e "/app/Yunzai-Bot/config/ICQQ.yaml" ]; then
             if [[ "ICQQ.yaml.bak" -nt "/app/Yunzai-Bot/config/ICQQ.yaml" ]]; then
                 cp ICQQ.yaml.bak /app/Yunzai-Bot/config/ICQQ.yaml
             fi
         fi
         if [[ -e "/app/Yunzai-Bot/config/config/ICQQ.yaml" ]]; then
-            if [[ ! -e "/app/Yunzai-Bot/config/ICQQ.yaml" ]]; then
-                echo -e "\n时雨：乐"
-                cp /app/Yunzai-Bot/config/config/ICQQ.yaml /app/Yunzai-Bot/config/ICQQ.yaml
-            else
-                if [[ "/app/Yunzai-Bot/config/config/ICQQ.yaml" -nt "/app/Yunzai-Bot/config/ICQQ.yaml" ]]; then
-                    cp /app/Yunzai-Bot/config/config/ICQQ.yaml /app/Yunzai-Bot/config/ICQQ.yaml
-                fi
-            fi
+            echo -e "\n时雨：乐"
         fi
         # 自用
         if [[ -e "/app/Yunzai-Bot/config/config/adapter/ICQQ.yaml" ]]; then
-            if [[ "/app/Yunzai-Bot/config/config/adapter/ICQQ.yaml" -nt "/app/Yunzai-Bot/config/ICQQ.yaml" ]]; then
+            if [ "/app/Yunzai-Bot/config/config/adapter/ICQQ.yaml" -nt "/app/Yunzai-Bot/config/ICQQ.yaml" ] || [ ! -e "/app/Yunzai-Bot/config/ICQQ.yaml" ]; then
                 cp app/Yunzai-Bot/config/config/adapter/ICQQ.yaml /app/Yunzai-Bot/config/ICQQ.yaml
             fi
         fi
@@ -111,31 +102,22 @@ if [[ -d $QQGUILD_PLUGIN_PATH"/.git" ]]; then
     else
         git pull origin main --allow-unrelated-histories
     fi
-    if [[ -e $QQGUILD_PLUGIN_PATH"/config.yaml" ]]; then
-        if [[ ! -e "/app/Yunzai-Bot/config/QQGuild.yaml" ]]; then
-            echo -e "检测到 QQGuild-Plugin 配置文件旧位置，复制到config/QQGuild.yaml\n"
-            cp config.yaml /app/Yunzai-Bot/config/QQGuild.yaml
-            mv config.yaml QQGuild.yaml.bak
-        fi
+    if [ -e $QQGUILD_PLUGIN_PATH"/config.yaml" ] && [ ! -e "/app/Yunzai-Bot/config/QQGuild.yaml" ]; then
+        echo -e "检测到 QQGuild-Plugin 配置文件旧位置，复制到config/QQGuild.yaml\n"
+        cp config.yaml /app/Yunzai-Bot/config/QQGuild.yaml
+        mv config.yaml QQGuild.yaml.bak
     else
-        if [[ -e "QQGuild.yaml.bak" ]]; then
+        if [ -e "QQGuild.yaml.bak" ] && [ -e "/app/Yunzai-Bot/config/QQGuild.yaml" ]; then
             if [[ "QQGuild.yaml.bak" -nt "/app/Yunzai-Bot/config/QQGuild.yaml" ]]; then
                 cp QQGuild.yaml.bak /app/Yunzai-Bot/config/QQGuild.yaml
             fi
         fi
         if [[ -e "/app/Yunzai-Bot/config/config/QQGuild.yaml" ]]; then
-            if [[ ! -e "/app/Yunzai-Bot/config/QQGuild.yaml" ]]; then
-                echo -e "\n时雨：乐"
-                cp /app/Yunzai-Bot/config/config/QQGuild.yaml /app/Yunzai-Bot/config/QQGuild.yaml
-            else
-                if [[ "/app/Yunzai-Bot/config/config/QQGuild.yaml" -nt "/app/Yunzai-Bot/config/QQGuild.yaml" ]]; then
-                    cp /app/Yunzai-Bot/config/config/QQGuild.yaml /app/Yunzai-Bot/config/QQGuild.yaml
-                fi
-            fi
+            echo -e "\n时雨：乐"
         fi
         # 自用
         if [[ -e "/app/Yunzai-Bot/config/config/adapter/QQGuild.yaml" ]]; then
-            if [[ "/app/Yunzai-Bot/config/config/adapter/QQGuild.yaml" -nt "/app/Yunzai-Bot/config/QQGuild.yaml" ]]; then
+            if [ "/app/Yunzai-Bot/config/config/adapter/QQGuild.yaml" -nt "/app/Yunzai-Bot/config/QQGuild.yaml" ] || [ ! -e "/app/Yunzai-Bot/config/QQGuild.yaml" ]; then
                 cp app/Yunzai-Bot/config/config/adapter/QQGuild.yaml /app/Yunzai-Bot/config/QQGuild.yaml
             fi
         fi
@@ -153,31 +135,22 @@ if [[ -d $TELEGRAM_PLUGIN_PATH"/.git" ]]; then
     else
         git pull origin main --allow-unrelated-histories
     fi
-    if [[ -e $TELEGRAM_PLUGIN_PATH"/config.yaml" ]]; then
-        if [[ ! -e "/app/Yunzai-Bot/config/Telegram.yaml" ]]; then
-            echo -e "检测到 Telegram-Plugin 配置文件旧位置，复制到config/Telegram.yaml\n"
-            cp config.yaml /app/Yunzai-Bot/config/Telegram.yaml
-            mv config.yaml Telegram.yaml.bak
-        fi
+    if [ -e $TELEGRAM_PLUGIN_PATH"/config.yaml" ] && [ ! -e "/app/Yunzai-Bot/config/Telegram.yaml" ]; then
+        echo -e "检测到 Telegram-Plugin 配置文件旧位置，复制到config/Telegram.yaml\n"
+        cp config.yaml /app/Yunzai-Bot/config/Telegram.yaml
+        mv config.yaml Telegram.yaml.bak
     else
-        if [[ -e "Telegram.yaml.bak" ]]; then
+        if [ -e "Telegram.yaml.bak" ] && [ -e "/app/Yunzai-Bot/config/Telegram.yaml" ]; then
             if [[ "Telegram.yaml.bak" -nt "/app/Yunzai-Bot/config/Telegram.yaml" ]]; then
                 cp Telegram.yaml.bak /app/Yunzai-Bot/config/Telegram.yaml
             fi
         fi
         if [[ -e "/app/Yunzai-Bot/config/config/Telegram.yaml" ]]; then
-            if [[ ! -e "/app/Yunzai-Bot/config/Telegram.yaml" ]]; then
-                echo -e "\n时雨：乐"
-                cp /app/Yunzai-Bot/config/config/Telegram.yaml /app/Yunzai-Bot/config/Telegram.yaml
-            else
-                if [[ "/app/Yunzai-Bot/config/config/Telegram.yaml" -nt "/app/Yunzai-Bot/config/Telegram.yaml" ]]; then
-                    cp /app/Yunzai-Bot/config/config/Telegram.yaml /app/Yunzai-Bot/config/Telegram.yaml
-                fi
-            fi
+            echo -e "\n时雨：乐"
         fi
         # 自用
         if [[ -e "/app/Yunzai-Bot/config/config/adapter/Telegram.yaml" ]]; then
-            if [[ "/app/Yunzai-Bot/config/config/adapter/Telegram.yaml" -nt "/app/Yunzai-Bot/config/Telegram.yaml" ]]; then
+            if [ "/app/Yunzai-Bot/config/config/adapter/Telegram.yaml" -nt "/app/Yunzai-Bot/config/Telegram.yaml" ] || [ ! -e "/app/Yunzai-Bot/config/Telegram.yaml" ]; then
                 cp app/Yunzai-Bot/config/config/adapter/Telegram.yaml /app/Yunzai-Bot/config/Telegram.yaml
             fi
         fi
@@ -195,31 +168,22 @@ if [[ -d $KOOK_PLUGIN_PATH"/.git" ]]; then
     else
         git pull origin main --allow-unrelated-histories
     fi
-    if [[ -e $KOOK_PLUGIN_PATH"/config.yaml" ]]; then
-        if [[ ! -e "/app/Yunzai-Bot/config/KOOK.yaml" ]]; then
-            echo -e "检测到 KOOK-Plugin 配置文件旧位置，复制到config/KOOK.yaml\n"
-            cp config.yaml /app/Yunzai-Bot/config/KOOK.yaml
-            mv config.yaml KOOK.yaml.bak
-        fi
+    if [ -e $KOOK_PLUGIN_PATH"/config.yaml" ] && [ ! -e "/app/Yunzai-Bot/config/KOOK.yaml" ]; then
+        echo -e "检测到 KOOK-Plugin 配置文件旧位置，复制到config/KOOK.yaml\n"
+        cp config.yaml /app/Yunzai-Bot/config/KOOK.yaml
+        mv config.yaml KOOK.yaml.bak
     else
-        if [[ -e "KOOK.yaml.bak" ]]; then
+        if [ -e "KOOK.yaml.bak" ] && [ -e "/app/Yunzai-Bot/config/KOOK.yaml" ]; then
             if [[ "KOOK.yaml.bak" -nt "/app/Yunzai-Bot/config/KOOK.yaml" ]]; then
                 cp KOOK.yaml.bak /app/Yunzai-Bot/config/KOOK.yaml
             fi
         fi
         if [[ -e "/app/Yunzai-Bot/config/config/KOOK.yaml" ]]; then
-            if [[ ! -e "/app/Yunzai-Bot/config/KOOK.yaml" ]]; then
-                echo -e "\n时雨：乐"
-                cp /app/Yunzai-Bot/config/config/KOOK.yaml /app/Yunzai-Bot/config/KOOK.yaml
-            else
-                if [[ "/app/Yunzai-Bot/config/config/KOOK.yaml" -nt "/app/Yunzai-Bot/config/KOOK.yaml" ]]; then
-                    cp /app/Yunzai-Bot/config/config/KOOK.yaml /app/Yunzai-Bot/config/KOOK.yaml
-                fi
-            fi
+            echo -e "\n时雨：乐"
         fi
         # 自用
         if [[ -e "/app/Yunzai-Bot/config/config/adapter/KOOK.yaml" ]]; then
-            if [[ "/app/Yunzai-Bot/config/config/adapter/KOOK.yaml" -nt "/app/Yunzai-Bot/config/KOOK.yaml" ]]; then
+            if [ "/app/Yunzai-Bot/config/config/adapter/KOOK.yaml" -nt "/app/Yunzai-Bot/config/KOOK.yaml" ] || [ ! -e "/app/Yunzai-Bot/config/KOOK.yaml" ]; then
                 cp app/Yunzai-Bot/config/config/adapter/KOOK.yaml /app/Yunzai-Bot/config/KOOK.yaml
             fi
         fi
@@ -237,31 +201,22 @@ if [[ -d $DISCORD_PLUGIN_PATH"/.git" ]]; then
     else
         git pull origin main --allow-unrelated-histories
     fi
-    if [[ -e $DISCORD_PLUGIN_PATH"/config.yaml" ]]; then
-        if [[ ! -e "/app/Yunzai-Bot/config/Discord.yaml" ]]; then
-            echo -e "检测到 Discord-Plugin 配置文件旧位置，复制到config/Discord.yaml\n"
-            cp config.yaml /app/Yunzai-Bot/config/Discord.yaml
-            mv config.yaml Discord.yaml.bak
-        fi
+    if [ -e $DISCORD_PLUGIN_PATH"/config.yaml" ] && [ ! -e "/app/Yunzai-Bot/config/Discord.yaml" ]; then
+        echo -e "检测到 Discord-Plugin 配置文件旧位置，复制到config/Discord.yaml\n"
+        cp config.yaml /app/Yunzai-Bot/config/Discord.yaml
+        mv config.yaml Discord.yaml.bak
     else
-        if [[ -e "Discord.yaml.bak" ]]; then
+        if [ -e "Discord.yaml.bak" ] && [ -e "/app/Yunzai-Bot/config/Discord.yaml" ]; then
             if [[ "Discord.yaml.bak" -nt "/app/Yunzai-Bot/config/Discord.yaml" ]]; then
                 cp Discord.yaml.bak /app/Yunzai-Bot/config/Discord.yaml
             fi
         fi
         if [[ -e "/app/Yunzai-Bot/config/config/Discord.yaml" ]]; then
-            if [[ ! -e "/app/Yunzai-Bot/config/Discord.yaml" ]]; then
-                echo -e "\n时雨：乐"
-                cp /app/Yunzai-Bot/config/config/Discord.yaml /app/Yunzai-Bot/config/Discord.yaml
-            else
-                if [[ "/app/Yunzai-Bot/config/config/Discord.yaml" -nt "/app/Yunzai-Bot/config/Discord.yaml" ]]; then
-                    cp /app/Yunzai-Bot/config/config/Discord.yaml /app/Yunzai-Bot/config/Discord.yaml
-                fi
-            fi
+            echo -e "\n时雨：乐"
         fi
         # 自用
-        if [[ -e "/app/Yunzai-Bot/config/config/adapter/Discord.yaml" ]]; then 
-            if [[ "/app/Yunzai-Bot/config/config/adapter/Discord.yaml" -nt "/app/Yunzai-Bot/config/Discord.yaml" ]]; then
+        if [[ -e "/app/Yunzai-Bot/config/config/adapter/Discord.yaml" ]]; then
+            if [ "/app/Yunzai-Bot/config/config/adapter/Discord.yaml" -nt "/app/Yunzai-Bot/config/Discord.yaml" ] || [ ! -e "/app/Yunzai-Bot/config/Discord.yaml" ]; then
                 cp app/Yunzai-Bot/config/config/adapter/Discord.yaml /app/Yunzai-Bot/config/Discord.yaml
             fi
         fi
